Added if condition on user option

+++ installer/tools/installer.ps1
@@ -35,27 +35,54 @@ if (-not $Url ) {
         Write-Output $versionNum
    }
     #https://github.com/Sensing-Dev/sensing-dev-installer/releases/download/v23.08.0-beta5/sensing-dev-installer-23.08.0-win64.zip
-    $Url = "https://github.com/Sensing-Dev/sensing-dev-installer/releases/download/${version}/sensing-dev-installer-${versionNum}-win64.zip"  
+    $zipUrl = "https://github.com/Sensing-Dev/sensing-dev-installer/releases/download/${version}/sensing-dev-installer-${versionNum}-win64.zip"
+    $msiUrl = "https://github.com/Sensing-Dev/sensing-dev-installer/releases/download/${version}/sensing-dev-installer-${versionNum}-win64.msi"
+
+    if ($user) {
+        # If a username is provided, get that user's LOCALAPPDATA path
+        $Url = "$zipUrl"
+    } else {
+        # If no username is provided, use the LOCALAPPDATA of the currently logged in user
+        $Url = "$msiUrl"
+    }   
+  
 }
 
-# Download ZIP to a temp location
-$tempZipPath = "$env:TEMP\$installerName.zip"
+if ($Url.EndsWith("zip")) {
+    # Download ZIP to a temp location
+    $tempZipPath = "$env:TEMP\$installerName.zip"
 
-Invoke-WebRequest -Uri $Url -OutFile $tempZipPath -Verbose
+    Invoke-WebRequest -Uri $Url -OutFile $tempZipPath -Verbose
+    # Unzip the file
+    Add-Type -AssemblyName System.IO.Compression.FileSystem
+    [System.IO.Compression.ZipFile]::ExtractToDirectory($tempZipPath, $installPath)
 
-# Unzip the file
-Add-Type -AssemblyName System.IO.Compression.FileSystem
-[System.IO.Compression.ZipFile]::ExtractToDirectory($tempZipPath, $installPath)
+    # Optionally delete the ZIP file after extraction
+    Remove-Item -Path $TargetFile -Force
 
-# Optionally delete the ZIP file after extraction
-Remove-Item -Path $TargetFile -Force
-
-# Check if the process started and finished successfully
-if ($?) {
-    Write-Host "The process ExtractToDirectory ran successfully."
-} else {
-    Write-Host "The process encountered an error."
+    # Check if the process started and finished successfully
+    if ($?) {
+        Write-Host "The process ExtractToDirectory ran successfully."
+    } else {
+        Write-Host "The process encountered an error."
+    }
 }
+
+if ($Url.EndsWith("msi")) {
+    # Download MSI to a temp location
+    $tempMsiPath = "$env:TEMP\$installerName.msi"
+    Invoke-WebRequest -Uri $Url -OutFile $tempMsiPath -Verbose
+
+    Start-Process -Wait -FilePath "msiexec.exe" -ArgumentList "/i `"$tempMsiPath`" TARGETDIR=`"$installPath`" INSTALL_ROOT=`"$installerName`" /qb /l*v $tempMsiPath\${installerName}_install.log" -Verb RunAs
+    # Check if the process started and finished successfully
+    if ($?) {
+        Write-Host "The process msiexec.exe ran successfully."
+    } else {
+        Write-Host "The process encountered an error."
+    }
+}
+
+
 
 # Run the .ps1 file from the installed package
 $ps1ScriptPath = Join-Path -Path $installPath -ChildPath $relativeScriptPath
